# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GLama documentation API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.88/api/v1",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получить все транзакции пользователя",
                "description": "Возможна фильтрация по полям type:deposit, deposit_yandex_account, removal, cashback. Дате date_start, date_end. Формат даты: 2024-05-18",
                "operationId": "378efe211748bf3cb7b82e9d83e71931",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All transactions"
                                        },
                                        "resource": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TransactionResponse"
                                                    }
                                                }
                                            },
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Some error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/refills": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получить все успешные транзакции на пополнение",
                "operationId": "dcd94e7d62188f842bbe2bf4b59130d2",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "resource": {
                                            "properties": {
                                                "amount": {
                                                    "type": "string",
                                                    "example": "5000.55"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Some error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/removals": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Получить все успешные транзакции на вывод денег",
                "operationId": "1b9ef46b0694eb42c6c4ee4719dd851d",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "resource": {
                                            "properties": {
                                                "amount": {
                                                    "type": "string",
                                                    "example": "5000.55"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Some error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pagination": {
                "title": "Пагинация",
                "properties": {
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://192.168.0.88/api/v1/tickets?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://192.168.0.88/api/v1/tickets?page=1"
                            },
                            "prev": {
                                "type": "null"
                            },
                            "next": {
                                "type": "null"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "links": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "url": {
                                            "type": "null"
                                        },
                                        "label": {
                                            "type": "string",
                                            "example": "&laquo; Назад"
                                        },
                                        "active": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "path": {
                                "type": "string",
                                "example": "http://192.168.0.88/api/v1/tickets"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 30
                            },
                            "to": {
                                "type": "integer",
                                "example": 1
                            },
                            "total": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "TransactionResponse": {
                "title": "Возвращаемая транзакция",
                "properties": {
                    "id": {
                        "title": "ID клиента (внутри базы)",
                        "description": "ID клиента (внутри базы)",
                        "format": "integer",
                        "example": "15"
                    },
                    "user_id": {
                        "title": "ID пользователя",
                        "description": "ID пользователя",
                        "format": "integer",
                        "example": "10"
                    },
                    "type": {
                        "title": "Тип транзакции",
                        "description": "Тип транзакции",
                        "format": "string",
                        "example": "deposit"
                    },
                    "status": {
                        "title": "Статус",
                        "description": "Статус",
                        "format": "string",
                        "example": "CONFIRMED"
                    },
                    "amount_deposit": {
                        "title": "Сумма без комиссии",
                        "description": "Сумма с комиссией",
                        "format": "integer",
                        "example": "30000"
                    },
                    "amount": {
                        "title": "Сумма с комиссией",
                        "description": "Сумма Сумма с комиссией",
                        "format": "integer",
                        "example": "30510"
                    },
                    "method_type": {
                        "title": "Тип пополнения",
                        "description": "Тип пополнения",
                        "format": "integer",
                        "example": "qr"
                    },
                    "created_at": {
                        "title": "Дата создания",
                        "description": "Дата создания",
                        "format": "date",
                        "example": "2024-04-02 09:53:15"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}